+++ assets/js/email.js
@@ -24,19 +24,4 @@ function emailCheck () {
     }
 }
 
-async function emailIApiCheck (email) {
-    const data = {
-        email: `${email}`
-    }
-    const res = await fetch(`https://bootcamp-api.codeit.kr/api/check-email
-    `, {
-        method: 'POST',
-        headers: {
-            'Content-Type' : 'application/json'
-        },
-        body: JSON.stringify(data)
-    });
-    return res;
-}
-
-export { emailEl, emailCheck, emailIApiCheck }
+export { emailEl, emailCheck }
